Add README and LICENSE
The first files we add to this repository are a README and the
license. We do this because we value documentation, and want to
make sure new users can get started as quickly and easy as
possible. In the README we'll document how the project works on a
high level, and how to install and deploy it. The license is there
to avoid any ambiguity and give users confidence that they are more
than welcome to use and extend our code.