Create and load default configuration file
We provide a configuration file with default values for all
configuration options. The purpose of this file is to:

a. provide documentation of all available parameters
b. allow users to quickly bootstrap an instance

Environment specific configuration files can be used to overwrite
the configuration and customize the behavior of the application.