Set up continuous integration
To make sure our application always complies with the high
standards we set for ourselves, we test and lint all the code we
publish to GitHub automatically via a continuous integration
pipeline.

For this project, we've chosen Codeship as the CI provider. Their
platform provides a lot of flexibility and great support for Docker
based builds, which is something that might come in handy later on
when adding more deployment options for customers.