Reorganize application and directory
The application and its directory have been reorganized to separate
the actual application code better from other project files. With
Codeship's configuration, Docker files and project documentation,
the root of the directory started to become too crowded to easily
distinguish between application and project files.

By moving the application into the `dec` package, it is clear that
the project root only has documentation and configuration files,
while the code sits in an isolated Python package.

While reorganizing the code, we also added setuptools into the mix
to make it easy to install the Python package.