Create directory for local DynamoDB
To enable users to develop and test their code with confidence, we
include instructions how to set up and run a local instance of
DynamoDB. The database and its libraries need to be downloaded, and
can be placed in this new directory.

Because the libraries is a product of Amazon, and because the
database may contain private information, we add the whole
directory to our .gitignore file. This way, we don't accidentally
share the intellectual property of a third party.