Add script to create the local Dynamo table
When the local DynamoDB is started for the first time, the user
needs to create the table. To make this a little bit easier, we
provide a script that does this. It assumes the user is using the
default configuration, which for development is very likely to be
the case.