Add configuration for DynamoDB
There are a few settings users might want to customize regarding
their deployment. For DynamoDB, the interesting settings are the
region, the endpoint (albeit almost identical with region) and the
table name. For development, we provide defaults for these values,
while for production, the user needs to customize these settings.