Create endpoint that stores webhooks in DynamoDB
The endpoint to store incoming webhooks in DynamoDB has been
created. If the incoming webhooks doesn't have data, HTTP 400 is
returned, and if saving the data fails, HTTP 500 is returned. If
parsing the request and storing the data succeeds, HTTP 200 gets
returned to the caller.

Before saving an item, a "dec_id" and a timestamp is added. While
not necessarily mandatory, it does make processing a little bit
easier in our experience.